API_AGENT_CODE/AutoTestAgent/requirements.txt
openai
mysql-connector-python
python-dotenv
requests